requirements.txt
asgiref==3.6.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
Django==4.1.7
django-ckeditor==6.5.1
django-embed-video==1.4.8
django-js-asset==2.0.0
idna==3.4
Pillow==9.4.0
psycopg2-binary==2.9.5
python-decouple==3.8
python-memcached==1.59
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
